chore!: move csv files into examples. put helper files into src/
